more messaging and changing names and docks
