+++ README.md
@@ -5,8 +5,8 @@
 ## Available Deployments Types
 | Deployment Type       | Stable | Link to Deployment                                    |
 |-----------------------|--------|-------------------------------------------------------|
-| On Prem - Simple      | ✔      | [Documentation](deployments/onprem_simple/README.md) |
-| On Prem - Distributed | ❌      | [Documentation](deployments/onprem_distributed/README.md) |
+| On Premise            | ✔      | [Documentation](deployments/onprem/README.md) |
+| On Premise - Distributed | ❌      | [Documentation](deployments/onprem_distributed/README.md) |
 | Amazon Web Service    | ❌      | [Documentation](deployments/aws/README.md) |
 | Microsoft Azure       | ❌      | [Documentation](deployments/azure/README.md) |
 | Google Cloud Platform | ❌      | [Documentation](deployments/gcp/README.md) |
@@ -21,7 +21,7 @@ Deployments are python modules that reside in `deployments`.  Each deployment mo
 #### Init
 This will generate a skeleton JSON configuration file for the deployment you choose to use.  It will output `deployment.json` to be used for specified deployment.
 
-Example: `python run.py init` or `python run.py init --type onprem_simple`
+Example: `python run.py init` or `python run.py init --type onprem`
 
 #### Deploy
 For a deployment to run, it must have the contents of `deployments/<deployment_type>/config-validator.json` satisfied.
@@ -36,7 +36,7 @@ If both the configuration file and the environment variables are set, the config
 | `--config <config-file>`                            | JSON formatted file required to deploy                                  |
 | `--type <type>`                                     | (Optional) Skip prompt and provide deployment type                      |
 
-Example: `python run.py deploy --type onprem_simple --config configuration.json`
+Example: `python run.py deploy --type onprem --config configuration.json`
 
 
 #### Remove Deployment
@@ -46,7 +46,7 @@ Tears down deployment provided.
 |-----------------------------------------------------|-------------------------------------------------------------------------|
 | `--type <type>`                                     | (Optional) Skip prompt and provide deployment type                      |
  
-Example: `python run.py undeploy --type onprem_simple`
+Example: `python run.py undeploy --type onprem`
 
 
 ## Database Support
